requirements.txt
fastapi==0.63.0
mangum==0.11.0
uvicorn==0.13.4
cx_oracle==8.2
python-dotenv
psycopg2-binary
pydantic
pytz
requests
python-multipart